requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
qrcode[pil]
